requirements.txt
pycryptodome
pyscreenshot
pynput